refactor: extract PDF front matter classification
